+++ bookmarks-components/src/index.js
@@ -1,0 +1,3 @@
+import Bookmark from './components/bookmark'
+
+export default Bookmark