+++ confiramtion-components/src/index.js
@@ -1,0 +1,3 @@
+import Confirmation from './components/confirmation-dialog'
+
+export default Confirmation