+++ header/src/components/header.js
@@ -1,21 +1,23 @@
 import Categories from './categories'
 import Channels from './channels'
 import Icons from './icons'
-import Logo from '../../static/twreporter-logo.svg'
 import LogoBright from '../../static/twreporter-logo-bright.svg'
+import LogoDark from '../../static/twreporter-logo.svg'
 import PropTypes from 'prop-types'
 import React from 'react'
+import SlideDownPanel from './header-slide-down-panel'
 import styled from 'styled-components'
 import { Link } from 'react-router'
 import { arrayToCssShorthand, screen } from 'shared/style-utils'
 import { colors } from 'shared/common-variables'
-import { HEADER_POSITION_UPON } from '../styles/constants'
+import { pageThemes } from 'shared/configs'
+import { selectBgColor } from '../styles/theme'
 
 const styles = {
   headerHeight: 109, // px
   headerHeightIndex: 62, // px
   topRowPadding: {
-    mobile: [34, 10, 35, 24], // px
+    mobile: [34, 24, 35, 24], // px
     tablet: [34, 20, 35, 35], // px
     desktop: [34, 58, 35, 70], // px
     index: {
@@ -38,7 +40,7 @@ const HeaderContainer = styled.div`
 `
 
 const TopRow = styled.div`
-  background-color: ${props => props.bgColor};
+  background-color: ${props => (props.isIndex ? colors.indexBodyBgWhite : selectBgColor(props.pageTheme))};
   height: ${props => (props.isIndex ? styles.headerHeightIndex : styles.headerHeight)}px;
 `
 
@@ -64,19 +66,68 @@ const TopRowContent = styled.div`
     max-width: ${props => (props.isIndex ? styles.topRowMaxWidth.desktop : styles.topRowMaxWidth.hd)}px;
   `}
   ${screen.hdAbove`
-    max-width: ${props => (props.headerPosition === HEADER_POSITION_UPON ? '100%' : `${styles.topRowMaxWidth.hd}px`)};
+    max-width: ${styles.topRowMaxWidth.hd}px;
   `}
   margin: 0 auto;
 `
 
+const HamburgerContainer = styled.div`
+  position: absolute;
+  ${screen.mobileOnly`
+    position: static;
+  `}
+`
+
+const HamburgerFrame = styled.div`
+  cursor: pointer;
+  display: none;
+  ${screen.mobileOnly`
+    display: initial;
+  `}
+`
+
+const Storke = styled.div`
+  width: 25px;
+  height: 4px;
+  margin-bottom: 5px;
+  background-color: ${colors.primaryColor};
+  border-radius: 10px;
+`
+
+const Hamburger = ({ onClick }) => (
+  <HamburgerContainer>
+    <HamburgerFrame onClick={onClick}>
+      <Storke />
+      <Storke />
+      <Storke />
+    </HamburgerFrame>
+  </HamburgerContainer>
+)
+
+Hamburger.propTypes = {
+  onClick: PropTypes.func.isRequired,
+}
+
 class Header extends React.PureComponent {
+  static _selectLogo(pageTheme) {
+    switch (pageTheme) {
+      case pageThemes.dark:
+        return <LogoDark onMouseDown={this._closeCategoriesMenu} />
+      case pageThemes.bright:
+      default:
+        return <LogoBright onMouseDown={this._closeCategoriesMenu} />
+    }
+  }
+
   constructor(props) {
     super(props)
     this.state = {
       categoriesIsOpen: false,
+      ifShowSlideInPanel: false,
     }
     this._closeCategoriesMenu = this._handleToggleCategoriesMenu.bind(this, 'close')
     this._handleToggleCategoriesMenu = this._handleToggleCategoriesMenu.bind(this)
+    this.handleOnHamburgerClick = this._handleOnHamburgerClick.bind(this)
   }
 
   componentWillReceiveProps(nextProps) {
@@ -95,60 +146,59 @@ class Header extends React.PureComponent {
     })
   }
 
-  _selectLogo(logoColor) {
-    switch (logoColor) {
-      case Header.logoColor.bright:
-        return <LogoBright onMouseDown={this._closeCategoriesMenu} />
-      case Header.logoColor.dark:
-      default:
-        return <Logo onMouseDown={this._closeCategoriesMenu} />
-    }
+  _handleOnHamburgerClick() {
+    this.setState({
+      ifShowSlideInPanel: !this.state.ifShowSlideInPanel,
+    })
   }
 
   render() {
-    const { bgColor, fontColor, logoColor, pathName, isIndex, headerPosition } = this.props
-    const { categoriesIsOpen } = this.state
+    const { pageTheme, pathName, isIndex, categoryId, signOutAction, ifAuthenticated } = this.props
+    const { categoriesIsOpen, ifShowSlideInPanel } = this.state
     return (
       <HeaderContainer>
-        <TopRow
-          bgColor={bgColor}
+        <SlideDownPanel
+          showUp={ifShowSlideInPanel}
           isIndex={isIndex}
-        >
-          <TopRowContent isIndex={isIndex} headerPosition={headerPosition} >
+          categoryId={categoryId}
+          ifAuthenticated={ifAuthenticated}
+          signOutAction={signOutAction}
+          handleOnHamburgerClick={this.handleOnHamburgerClick}
+        />
+        <TopRow pageTheme={pageTheme} isIndex={isIndex}>
+          <TopRowContent isIndex={isIndex}>
             <Link to="/">
-              {this._selectLogo(logoColor)}
+              {Header._selectLogo(pageTheme)}
             </Link>
-            <Icons />
+            <Icons
+              pageTheme={pageTheme}
+              ifAuthenticated={ifAuthenticated}
+              signOutAction={signOutAction}
+            />
+            <Hamburger onClick={this.handleOnHamburgerClick} />
           </TopRowContent>
         </TopRow>
-        {isIndex ? null : <Channels handleToggleCategoriesMenu={this._handleToggleCategoriesMenu} fontColor={fontColor} pathName={pathName} categoriesIsOpen={categoriesIsOpen} headerPosition={headerPosition} />}
-        {isIndex ? null : <Categories categoriesIsOpen={categoriesIsOpen} handleToggleCategoriesMenu={this._handleToggleCategoriesMenu} bgColor={bgColor} />}
+        {isIndex ? null : <Channels handleToggleCategoriesMenu={this._handleToggleCategoriesMenu} pageTheme={pageTheme} pathName={pathName} categoriesIsOpen={categoriesIsOpen} />}
+        {isIndex ? null : <Categories categoriesIsOpen={categoriesIsOpen} handleToggleCategoriesMenu={this._handleToggleCategoriesMenu} pageTheme={pageTheme} />}
       </HeaderContainer>
     )
   }
 }
 
-Header.logoColor = {
-  dark: 'dark',
-  bright: 'bright',
-}
-
 Header.propTypes = {
-  bgColor: PropTypes.string,
-  fontColor: PropTypes.string,
-  logoColor: PropTypes.string,
+  pageTheme: PropTypes.string,
   pathName: PropTypes.string,
   isIndex: PropTypes.bool,
-  headerPosition: PropTypes.string,
+  ifAuthenticated: PropTypes.bool.isRequired,
+  signOutAction: PropTypes.func.isRequired,
+  categoryId: PropTypes.string,
 }
 
 Header.defaultProps = {
-  bgColor: '',
-  fontColor: colors.black,
-  logoColor: Header.logoColor.dark,
+  pageTheme: pageThemes.bright,
   isIndex: false,
   pathName: '',
-  headerPosition: '',
+  categoryId: '',
 }
 
 export default Header