+++ header/src/components/icons.js
@@ -1,14 +1,16 @@
 import React from 'react'
-import styled from 'styled-components'
+import PropTypes from 'prop-types'
 import SearchBox from './search-box'
 import SearchIcon from '../../static/search-icon.svg'
+import styled from 'styled-components'
 import { screen } from 'shared/style-utils'
 import { colors, fonts } from 'shared/common-variables'
-import { searchConfigs } from 'shared/configs'
+import { searchConfigs, memberConfigs, bookmarkConfigs } from 'shared/configs'
 import { Link } from 'react-router'
 
-// import MemberIcon from '../../static/member-icon.svg'
-// import BookmarkListIcon from '../../static/bookmark-list-icon.svg'
+import SignInIcon from '../../static/service-icons/member-icon.svg'
+import SignOutIcon from '../../static/service-icons/signout.svg'
+import BookmarkListIcon from '../../static/bookmark-list-icon.svg'
 // import DonateIcon from '../../static/donate-icon.svg'
 
 const styles = {
@@ -18,11 +20,17 @@ const styles = {
 /* Icon alt text takes 2 fullwidth characters specifically */
 const ICON_ALT_TEXT = {
   SEARCH: '搜尋',
+  SIGN_OUT: '登出',
+  MEMBER: '會員',
+  BOOKMARK: '書籤',
 }
 
 const IconsContainer = styled.div`
   position: relative;
   display: table;
+  ${screen.mobileOnly`
+    display: none;
+  `}
 `
 
 const IconContainer = styled.div`
@@ -117,7 +125,22 @@ class Icons extends React.PureComponent {
     })
   }
   render() {
+    const { ifAuthenticated } = this.props
     const { isSearchOpened } = this.state
+    const Member = (() => {
+      const { signOutAction } = this.props
+      return (
+        <Link
+          to={`/${memberConfigs.path}`}
+          onClick={() => {
+            signOutAction()
+          }}
+        >
+          {ifAuthenticated ? <SignOutIcon /> : <SignInIcon />}
+          <span>{ifAuthenticated ? `${ICON_ALT_TEXT.SIGN_OUT}` : `${ICON_ALT_TEXT.MEMBER}` }</span>
+        </Link>
+      )
+    })()
     return (
       <IconsContainer>
         <DisplayOnDesktop
@@ -136,10 +159,23 @@ class Icons extends React.PureComponent {
             <SearchIcon />
           </Link>
         </HideOnDesktop>
+        <IconContainer>
+          <Link to={`/${bookmarkConfigs.path}`}>
+            <BookmarkListIcon />
+            <span>{ICON_ALT_TEXT.BOOKMARK}</span>
+          </Link>
+        </IconContainer>
+        <IconContainer>
+          {Member}
+        </IconContainer>
       </IconsContainer>
     )
   }
 }
 
+Icons.propTypes = {
+  ifAuthenticated: PropTypes.bool.isRequired,
+  signOutAction: PropTypes.func.isRequired,
+}
 
 export default Icons