+++ index-page/src/constants/strings.js
@@ -2,7 +2,7 @@ const strings = {
   editorPick: '編輯精選',
   latestTopic: '最新專題',
   review: '評論',
-  category: '分類',
+  category: '議題',
   topic: '專題',
   photography: '攝影',
   infographic: '多媒體',