+++ main.js
@@ -1,14 +1,14 @@
-import { Cinemagraph, CinemagraphHelper } from './cinemagraph'
+import Bookmark from './bookmarks-components/src'
+import Confirmation from './confiramtion-components/src'
+import FadeText from './fade-text'
 import FetchingWrapper from './shared/components/is-fetching-wrapper'
 import Footer from './footer'
 import Header from './header'
 import IndexPageComposite from './index-page'
-import ListingComposite from './listing-page'
+import { TopicsList, List } from './listing-page'
 import Pagination from './pagination'
-import FadeText from './fade-text'
+import { Cinemagraph, CinemagraphHelper } from './cinemagraph'
 import { FullpageSlides, FullpageSlidesConstants } from './fullpage-slides'
-
-const { TopicsList, List } = ListingComposite
 
 const Wrapper = {
   FetchingWrapper,
@@ -21,9 +21,11 @@ export default {
   FullpageSlides,
   FullpageSlidesConstants,
   Header,
+  Bookmark,
   IndexPageComposite,
   List,
   Pagination,
+  Confirmation,
   FadeText,
   TopicsList,
   Wrapper,
@@ -36,9 +38,11 @@ export {
   FullpageSlides,
   FullpageSlidesConstants,
   Header,
+  Bookmark,
   IndexPageComposite,
   List,
   Pagination,
+  Confirmation,
   FadeText,
   TopicsList,
   Wrapper,